+++ CHANGELOG.md
@@ -5,9 +5,11 @@ CHANGELOG
 
   * Extract process run management to ProcessRunner.
   * Add support for process listeners.
+  * Provides bundled DebugListener
   * BC break : ProcessRunnerInterface::run now takes an SplObjectStorage containing
     listeners as second argument.
-  * Provides bundled DebugListener
+  * BC break : BinaryInterface no longer implements LoggerAwareInterface
+    as it is now supported by ProcessRunner.
 
 * 1.3.4 (2013-04-26)